created the user.php class
This class handles the getting and setting user information. As well as
adding a user. I also created a method in the controller and javascript
files to cater for logginig in.
I also created separate pages for each usertype. That is for admin,
supervisor and nurse. For each implementation it has to be done with the
respective usertype pages